requirements.txt
Flask==2.3.3
Flask_Cors==3.0.10
flask_sqlalchemy==3.0.3
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.6
flask_jwt_extended==4.4.4
flask_socketio==5.3.3
Flask-Migrate==4.0.4
pymysql==1.0.2
alembic==1.10.3
jsonschema~=4.17.3
Flask-Bcrypt~=1.0.1
PyJWT~=2.7.0
flask_jwt_extended==4.4.4
python-dotenv==1.0.0
Flask-Pydantic==0.11.0
pydantic~=1.10.11
dependency-injector==4.41.0